SpCrud folder structure
