+++ packages/vue/src/components/SpCrudUpdate/index.ts
@@ -1,0 +1,13 @@
+import { App as Application } from 'vue'
+
+import { registerComponent } from './../../utils/plugins/index'
+// @ts-ignore
+import C from './SpCrudUpdate.vue'
+
+export const Plugin = {
+  install(vue: Application): void {
+    registerComponent(vue, C)
+  }
+}
+
+export default C
