+++ packages/vue/src/components/index.ts
@@ -1,14 +1,18 @@
 export { default as SpAcc } from './SpAcc'
 export { default as SpAmountSelect } from './SpAmountSelect'
-export { default as SpAssets } from './SpAssets'
 export { default as SpButton } from './SpButton'
 export { default as SpCardNew } from './SpCard'
 export { default as SpClipboard } from './SpClipboard'
-export { default as SpDropdown } from './SpDropdown'
 export { default as SpLinkIcon } from './SpLinkIcon'
 export { default as SpModal } from './SpModal'
 export { default as SpNavbar } from './SpNavbar'
-export { default as SpSpacer } from './SpSpacer'
 export { default as SpTheme } from './SpTheme'
-export { default as SpTx } from './SpTx'
+export { default as SpSpacer } from './SpSpacer'
 export { default as SpTypography } from './SpTypography'
+export { default as SpDropdown } from './SpDropdown'
+export { default as SpTx } from './SpTx'
+export { default as SpCrudCreate } from './SpCrudCreate'
+export { default as SpCrudRead } from './SpCrudRead'
+export { default as SpCrudUpdate } from './SpCrudUpdate'
+export { default as SpCrudDelete } from './SpCrudDelete'
+export { default as SpCrud } from './SpCrud'
